+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/api/HubdleConfigDocumentationApiExtension.kt
@@ -55,3 +55,6 @@ constructor(
 
 internal val HubdleEnableableExtension.hubdleApi: HubdleConfigDocumentationApiExtension
     get() = getHubdleExtension()
+
+internal val Project.hubdleApi: HubdleConfigDocumentationApiExtension
+    get() = getHubdleExtension()