+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/site/HubdleConfigDocumentationSiteExtension.kt
@@ -71,3 +71,6 @@ constructor(
 
 internal val HubdleEnableableExtension.hubdleSite: HubdleConfigDocumentationSiteExtension
     get() = getHubdleExtension()
+
+internal val Project.hubdleSite: HubdleConfigDocumentationSiteExtension
+    get() = getHubdleExtension()