+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/install/pre/commits/tasks/InstallCheckAnalysisPreCommitTask.kt
@@ -1,6 +1,5 @@
 package com.javiersc.hubdle.project.extensions.config.install.pre.commits.tasks
 
-import com.javiersc.gradle.tasks.extensions.namedLazily
 import javax.inject.Inject
 import org.gradle.api.Project
 import org.gradle.api.file.ProjectLayout
@@ -8,6 +7,7 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.TaskAction
+import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
 
@@ -32,19 +32,19 @@ constructor(
     }
 
     public companion object {
-        public const val name: String = "installCheckAnalysisPreCommit"
+        public const val NAME: String = "installCheckAnalysisPreCommit"
 
         internal fun register(project: Project) {
             val installCheckAnalysisPreCommitTask =
-                project.tasks.register<InstallCheckAnalysisPreCommitTask>(name)
+                project.tasks.register<InstallCheckAnalysisPreCommitTask>(NAME)
 
             installCheckAnalysisPreCommitTask.configure { task ->
                 task.finalizedBy(WriteFilePreCommitTask.getTask(project))
             }
 
-            project.tasks
-                .namedLazily<InstallPreCommitTask>(InstallPreCommitTask.name)
-                .configureEach { task -> task.dependsOn(installCheckAnalysisPreCommitTask) }
+            project.tasks.named<InstallPreCommitTask>(InstallPreCommitTask.NAME).configure { task ->
+                task.dependsOn(installCheckAnalysisPreCommitTask)
+            }
         }
     }
 }