+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/tasks/HubdleTask.kt
@@ -1,0 +1,3 @@
+package com.javiersc.hubdle.project.tasks
+
+public interface HubdleTask