+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/settings/tasks/HubdleTask.kt
@@ -1,0 +1,1 @@
+package com.javiersc.hubdle.settings.tasks